profile: add util header to separate functions to browse profiles
