+++ src/AvTranscoder/profile/ProfileLoader.hpp
@@ -4,7 +4,6 @@
 #include <AvTranscoder/common.hpp>
 
 #include <string>
-#include <cstring>
 #include <vector>
 #include <map>
 
@@ -72,11 +71,6 @@ private:
 private:
 	Profiles _profiles;
 };
-
-#ifndef SWIG
-void split( std::vector< std::string >& splitString, const std::string& inputString, const std::string& splitChars = ";" );
-int getFilesInDir( const std::string& dir, std::vector< std::string >& files );
-#endif
 
 }
 #endif
